requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
colorama==0.4.6
fastapi==0.128.4
h11==0.16.0
idna==3.11
ImageIO==2.37.2
joblib==1.5.3
lazy_loader==0.4
networkx==3.6.1
numpy==2.4.2
opencv-python-headless==4.13.0.92
packaging==26.0
pillow==12.1.0
pydantic==2.12.5
pydantic_core==2.41.5
python-multipart==0.0.22
scikit-image==0.26.0
scikit-learn==1.6.1
scipy==1.17.0
starlette==0.52.1
threadpoolctl==3.6.0
tifffile==2026.1.28
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
